banking_agent_demo/requirements.txt
anthropic
python-dotenv
pydantic
streamlit
